+++ NEWS.md
@@ -1,9 +1,19 @@
 # Spectra 1.11
 
-## Changes in 1.11.7
+## Changes in 1.11.8
 
 - Add `filterPrecursorPeaks` to allow filtering peaks within each spectrum
   with m/z values relative to the precursor m/z of the spectrum.
+
+## Changes in 1.11.7
+
+- Fix in `compareSpectra` to also pass parameters `ppm` and `tolerance` to
+  the peak similarity calculation functions `FUN`: this allows to use custom
+  similarity function with integrated mapping of peaks.
+- Add `joinPeaksNone` to skip the peak matching in `compareSpectra` if the
+  similarity scoring function performs its own peak matching.
+- Only use parallel processing in `setBackend,Spectra` if both backends support
+  it.
 
 ## Changes in 1.11.6